Rename plugins in /esp32
