+++ src/modules/esp32/RangeTestModule.cpp
@@ -1,4 +1,4 @@
-#include "RangeTestPlugin.h"
+#include "RangeTestModule.h"
 #include "MeshService.h"
 #include "NodeDB.h"
 #include "PowerFSM.h"
