+++ src/modules/esp32/StoreForwardModule.cpp
@@ -1,4 +1,4 @@
-#include "StoreForwardPlugin.h"
+#include "StoreForwardModule.h"
 #include "MeshService.h"
 #include "NodeDB.h"
 #include "RTC.h"
